requirements.txt
numpy==1.25.2
scipy==1.11.2
opencv-python==4.9.0.80
networkx==3.1  # required by torch
torch==2.0.1
pillow==10.0.0
certifi==2023.7.22  # required by requests
charset-normalizer==3.2.0  # required by requests
idna==3.4  # required by requests
urllib3==2.0.4  # required by requests
requests==2.31.0
torchvision==0.15.2
joblib==1.3.2
threadpoolctl==3.4.0
scikit-learn==1.4.1.post1
contourpy==1.1.0  # required by matplotlib
cycler==0.12.1  # required by matplotlib
fonttools==4.42.1  # required by matplotlib
kiwisolver==1.4.5  # required by matplotlib
packaging==23.1  #required by matplotlib
pyparsing==3.0.9  # required by matplotlib
six==1.16.0  # required by python-dateutil
python-dateutil==2.8.2  # required by matplotlib
pandas==2.1.1
matplotlib==3.7.2
seaborn==0.13.2
jsonpickle==3.0.3